requirements.txt
faiss-cpu==1.8.0.post1
ollama==0.3.1
openai==1.40.6
pypdf==4.3.1
python-dotenv==1.0.1
sentence-transformers==3.0.1
streamlit==1.38.0
tiktoken==0.7.0